Clean up a bunch of now-unused code
