+++ src/Client/src/app/projects/projects.component.html
@@ -1,1 +1,6 @@
-<p>projects works!</p>
+<app-data-table [dataSource]="dataSource"
+[columns]="columns"
+[pageSize]="3"
+[pageSizeOptions]="[1,2,3,5,7,10]"
+(itemSelected)="itemSelected($event)"
+></app-data-table>
